changes ivan jorge

+++ src/Controller/EmpresaController.php
@@ -39,9 +39,11 @@ class EmpresaController extends AbstractController
             ->getRepository(Empresa::class)
             ->findAll();
 
-       
-
-        return $this->render('empresa/list.html.twig', ['empresas' => $empresas]);
+        if(isset($_REQUEST['mensaje']) && $_REQUEST['mensaje'] == 'error'){
+                return $this->render('empresa/list.html.twig', ['empresas' => $empresas, 'mensaje' => "Error , no se ha podido eliminar esta corporación. Contiene una o varias empresas, por favor primero elimina esas empresas"]);
+        } else {
+                return $this->render('empresa/list.html.twig', ['empresas' => $empresas, 'mensaje' => " "]);
+        }
     }
 
     /**
@@ -174,21 +176,30 @@ class EmpresaController extends AbstractController
      */
     public function delete($id, Request $request)
     {
+        $mensajeErrorFK="";
         $empresas = $this->getDoctrine()
             ->getRepository(Empresa::class)
             ->find($id);
+        
+        $centros=$empresas->getArrayCentros();   
+        if($centros->isEmpty()){ 
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($empresas);
+            $entityManager->flush();
+            $nomEmpresas = $empresas->getNombre();
+            $this->addFlash(
+                'notice',
+                'Empresa '.$nomEmpresas.' eliminada!'
+            );
+        } else {
+            $mensajeErrorFK="error";
+        }
+        
 
-        $entityManager = $this->getDoctrine()->getManager();
-        $nomEmpresas = $empresas->getNombre();
-        $entityManager->remove($empresas);
-        $entityManager->flush();
-
-        $this->addFlash(
-            'notice',
-            'Empresa '.$nomEmpresas.' eliminada!'
+        return $this->redirectToRoute('empresas_list', array(
+            'mensaje'=>$mensajeErrorFK,
+            )
         );
-
-        return $this->redirectToRoute('empresas_list');
     }
